Moving sprites to their own file inside controllers and created Level file
