+++ src/game_modules/controllers/__init__.py
@@ -1,0 +1,4 @@
+
+'''
+    Controllers folder is where the controllers scripts will be.
+'''
